+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests for dev-site v6
 
 4.8.1 - (July 23, 2019)
 ------------------