+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+
+4.8.1 - (July 23, 2019)
+------------------
+* a completely unremarkable update because lerna be lerna
+
+4.8.0 - (July 23, 2019)
+------------------
 ### Changed
 * Normalized terra-visually-hidden-text dependency version to match other terra-framework packages